Example files, functions in separate sub-module
